UPSTREAM:DP:wifi:ath12k: Write hw_peer_id while updating HAL_TCL_DATA_CMD_INFO4_SEARCH_INDEX
In case of 4-address multicast packet, it is expected to write
ast_idx(hw_peer_id) of peer before queueing the packet to TCL.
Currently it is filled with peer_id which will not be same
in all the time. This would trigger some unknown behavior in
the traffic running system.

Fixes: I30d21218f1 ("UPSTREAM:CP:DP: ath12k: Add wds 4addr mode support for AP/STA")

Patch-dependency: 0123-0002-UPSTREAM-DP-ath12k-Add-4addr-tx-multicast-packet-sup.patch

Change-Id: I9a82c25110aa1365b0d2f8d7dc44188b334eff17
Signed-off-by: Tamizh Chelvam Raja <tamizhr@qti.qualcomm.com>