QSDK:PLATFORM:ath12k:fix coredump collection and add transfer size tracking
Currently during pci coredump, ELF header size is passed to
dev_coredumpm() instead of the total dump size. In of the dump
collection scenarios the dump size is less the exepected size

Fix this by storing the total dump size during ELF build and pass it to
the dev_coredumpm() framework.
Additionally, add a print to track the number of bytes transferred during
read.

Patch-dependency: None
Change-Id: I70672d0a51e379733f6ac7c5e111af09f9ceae3c
Signed-off-by: Narendhar Reddy <nare@qti.qualcomm.com>