UPSTREAM:DP:wifi:ath12k: Remove unused arsta in Rx monitor path
In Rx monitor path, there is API call to ath12k_peer_get_link_sta()
to fetch arsta. This API call requires RCU lock to be taken, which
is missing and this can lead to use-afer-free scenario.

This arsta is used in this path just for the NULL check and therefore
this arsta fetch and check can be safely removed and hence RCU lock
also shall not be required.

Patch-dependency: I57f727eb5ae0 ("Upstream:DP: Add logic for PPDU based reaping for monitor destination ring")
                  I4da8140f41d5 ("UPSTREAM CP DP: wifi: ath12k: Add telemetry stats for QCA Energy Service")

Change-Id: Ifcb9e72830101cef6a3340314eb111c59f62bd30
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>