UPSTREAM:PLATFORM:wifi:ath12k: Handle rmmod during SSR scenario
Currently, when rmmod is performed during SSR recovery, rddm download
is issuing BUG_ON() even though the minidump is already uploaded.

Ideally, if the FW assert happens after rmmod is issued or as part of
the rmmod sequence, the expectation is to capture the FW dumps and
then go for a BUG_ON(). So, add changes to skip dump collection if
core_reset is already queued, as it will take care of uploading the
dump and powering down the Q6.

Patch-dependency: Iac6f78468d3 (UPSTREAM:PLATFORM:wifi: ath12k: Skip core reset in case of FW crash)

Change-Id: I2fd931a1cbcd36d04e61a6b0b2c00bb078aad836
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>