UPSTREAM:CP:wifi:ath12k: Do dp_peer_cleanup on Mode-0 recovery.
Currently as part of Mode-0 recovery flow though the low ack is
triggered for the connected stations for mesh DEL_STATION is
not received because of which local->sta_list will have this
mesh entry and upon restart_hw() mac80211 will try to re add
this entries.

But in post_reconfigure_recovery() dp_peer_cleanup() is not
being done for Mode-0 recovery becase of which the mesh dp_peer
will be present and when mac80211 tries to re add it peer_create
will fail which throws a warning.

Ideally ath12k should have removed its entry and then requested
for restart so that mac80211 can re add it but because for Mode-0
ath12k_mac_dp_peer_cleanup() was skipped this issue is seen.

Fix this by doing dp_peer_cleanup for both Mode-0, Mode-1 and
skipping it for Mode-2.

Fixes: I5f4dfc7d61a (wifi:ath12k: Add Mode2 recovery support for MLO)

Patch-dependency: I5f4dfc7d61a (wifi:ath12k: Add Mode2 recovery support for MLO)

Change-Id: Ia46dc96c0e934c41ea281e1f5ed4b193befb9167
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>