UPSTREAM:CP:wifi:ath12k:Support 512 max client
The maximum supported clients per radio is now increased to 512. A module
parameter remains available to reduce this limit for testing purposes.
Additionally, in QCN5424, firmware can now support 512 clients, hence
changing the same in host as well.

For MLO, the maximum supported client count is determined by the minimum
of the maximum supported MLO clients per SoC in a DUT.

Patch-dependency: none

Change-Id: I8b7cadaeb66cd261c2bb2939bb88dfb1347e9285
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>