UPSTREAM:PLATFORM:wifi:ath12k: Populate DDR range in host capabilities
As per new request, host is required to share DDR range information
via QMI to FW which will use it to configure the outbound (OB) blocker.
This will be used to address issues due to invalid memory access and
access violations due to permission issues.

Fetch memory node from the dts and store the start and end address
in mhi_ctrl during MHI registration. These values are then used
to populate the start address and size of the DDR range in the host
capabilities QMI message to FW for the outbound blocker.

Patch-dependency: None

Change-Id: I9473d3b1d5688556b816b03857c61e1ba64962a4
Signed-off-by: Sajini R <sajinir@qti.qualcomm.com>