UPSTREAM:CP:wifi:ath12k Store chip frequency range in ar
Hostapd crash is observed on a AP with a 3 link (2, 5 and 6 GHz) AP MLD,
when country code "IN" is configured with the following sequence of
commands:

1. hostapd_cli -i wlan1 -l 0 disable
2. hostapd_cli -i wlan1 -l 0 set country_code 'IN'
3. hostapd_cli -i wlan1 -l 0 enable

6 GHz is not supported on country code "IN". The start and end frequency
of the 6G hardware is filled with zeroes to the Hostapd, in the
NL80211_CMD_GET_WIPHY during the country configuration in the Hostapd.
In ath12k_regd_update, the start and end frequencies of the radio
are filled as an intersection of the radio frequency range and the
regulatory frequency range. In case of country "IN", the regulatory
frequency range has the start and end frequency as zeroes.

Fix this issue by storing the radio's start and end frequencies in the ar
instead of intersecting them with the regulatory start and end freq.
Implement this logic in ath12k_mac_setup_channels_rates (country
configuration during boot up) and ath12k_regd_update.

Tag: REGULATORY
Fixes: Ia00a4ebb85c('QSDK:CP: wifi: ath12k: Port changes from Upstream')
Patch-dependency: None

Change-Id: I7288a6b118967949806a34fcd608a84bf956fdcc
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>