QSDK:PLATFORM:wifi:ath12k: Add probe order support for SOC Initialisation
Introduce probe_order configuration in ath12k to control the initialization sequence of SoCs. The probe_order value must be specified in hexadecimal format, where each digit represents the position of a SoC in the probe sequence.

For example:
- Default order defined in dts: Soc0 -> Soc2 -> Soc1
- probe_order = 0x213
                  |||
                  |||
                  |||----->Soc0->Order 3
                  ||
                  ||------>Soc2->Order 1
                  |
                  |------->Soc1->Order 2

results in: Soc2 -> Soc1 -> Soc0

If probe_order is 0 or not specified, the default order will be applied. This feature provides flexibility in managing platform device dependencies during driver bring-up.

Patch-dependency: none

Change-Id: I9246567227261a3f8a23fe40c4591341d2238e3c
Signed-off-by: Sham Muthayyan <smuthayy@qti.qualcomm.com>