UPSTREAM: DP: WiFi: ATH12k: Fix race condition with link_id in Tx completion
The `peer->stats_link_id` field was used to cache the `link_id`
derived from the HTT Tx completion. This cached value was then used
by subsequent functions to update peer statistics and to look up
link-specific peers.

This caching mechanism can lead to a race condition in Multi-Link
Operation (MLO) scenarios. If Tx completions for different links
arrive concurrently for the same peer, they could overwrite the
`stats_link_id` value. This would cause statistics to be updated
for the wrong link or lead to incorrect link-peer lookups.

To fix this, the `stats_link_id` field is removed from the
`ath12k_dp_peer` struct. The `link_id` is now fetched locally within
the Tx completion handler and passed as a direct argument to the
functions that require it. This ensures that the correct `link_id`
is used for each completion, eliminating the race condition.

Patch-dependency:
0426-0005-UPSTREAM-DP-wifi-ath12k-Telemetry-counters-in-Tx-com.patch

Change-Id: Id907cdf0df34b38a9c9a059a1799bb886d56111e
Signed-off-by: Kenvish Butani <kbutani@qti.qualcomm.com>