UPSTREAM:CP:wifi:ath12k: Add sanity check before stopping the bridge
When the last normal vdev is stopped, all the bridge vdevs will
be stopped. Similarly when the last normal vdev is deleted, all
the bridge vdevs will be deleted. In issue scenario, the last
normal vdev is getting stopped but not deleted. Following this
another vdev is getting added and removed. During its removal,
in vdev stop process once all the normal vdevs are stopped,
the bridge vdev stop is also invoked which should not be done
in this case.

Add sanity check to avoid stopping bridge vdev when it is not
yet started.

Patch-dependency: Fixes: I526e7fa10c6 ("UPSTREAM: CP: wifi: ath12k: Rewrite bridge vdev removal logic")

Change-Id: I63f01fee60eb4c703ff7bbd42f529304dbf818da
Signed-off-by: Nagarajan Maran <nmaran@qti.qualcomm.com>